+++ src/helpers/filter-notification.js
@@ -6,14 +6,16 @@ import _ from "lodash";
  * This version should filter out all users if the `synchronized_segments`
  * setting is empty
  * @param  {Object} ctx The Context Object
- * @param  {Object} user Hull user object
+ * @param  {Object} notification Hull user:update notification
  * @param  {String} fieldName the name of settings name
  * @return {Boolean}
  */
-export default function filterUserSegments(ctx, user, fieldName = "synchronized_segments") {
+export default function filterNotification(ctx, notification, fieldName = "synchronized_segments") {
   if (!_.has(ctx.ship.private_settings, fieldName)) {
     return true;
   }
   const filterSegmentIds = _.get(ctx.ship.private_settings, fieldName, []);
-  return _.intersection(filterSegmentIds, user.segment_ids).length > 0;
+
+  const segments = _.uniq(_.concat(notification.segments || [], _.get(notification, "changes.segments.left", [])));
+  return _.intersection(filterSegmentIds, segments.map(s => s.id)).length > 0;
 }