+++ src/helpers/index.js
@@ -1,4 +1,4 @@
-export filterUserSegments from "./filter-user-segments";
+export filterNotification from "./filter-notification";
 export requestExtract from "./request-extract";
-export setUserSegments from "./set-user-segments";
+// export setUserSegments from "./set-user-segments";
 export updateSettings from "./update-settings";