+++ src/utils/batch-handler.js
@@ -1,4 +1,5 @@
 import { Router } from "express";
+import _ from "lodash";
 
 import { group } from "../trait";
 import responseMiddleware from "./response-middleware";
@@ -11,18 +12,26 @@ export default function batchHandler(handler, { batchSize = 100, groupTraits = f
   const router = Router();
   router.use(requireHullMiddleware());
   router.post("/", (req, res, next) => {
-    const { client, helpers } = req.hull;
+    const { client, segments } = req.hull;
 
     return client.utils.extract.handle({
       body: req.body,
       batchSize,
       handler: (users) => {
         const segmentId = req.query.segment_id || null;
-        users = users.map(u => helpers.setUserSegments({ add_segment_ids: [segmentId] }, u));
         if (groupTraits) {
           users = users.map(u => group(u));
         }
-        return handler(req.hull, users);
+
+        const messages = users.map((u) => {
+          const segmentIds = _.uniq(_.concat(u.segment_ids || [], [segmentId]));
+          return {
+            user: u,
+            segments: segmentIds.map(id => _.find(segments, { id }))
+          };
+        });
+
+        return handler(req.hull, messages, { query: req.query, body: req.body });
       }
     }).then(next, next);
   });